+++ neil_test.sh
@@ -6,5 +6,5 @@ source source_me.sh && ./run_pipeline.py -i scantrons/ \
   -k scantrons/804D5A50-key.jpg \
   -o output/ -d
 
-sleep 1
-output/*_debug.png
+sleep 0.1
+open output/*_debug.png